+++ engine/include/cubos/engine/transform/scale.hpp
@@ -2,11 +2,11 @@
 
 #include <glm/glm.hpp>
 
-namespace cubos::engine::ecs
+namespace cubos::engine
 {
     /// Assigns a uniform scale to an entity.
     struct [[cubos::component("cubos/scale", VecStorage)]] Scale
     {
         float factor; ///< The scale factor of the entity.
     };
-} // namespace cubos::engine::ecs
+} // namespace cubos::engine
