+++ engine/src/cubos/engine/transform/plugin.cpp
@@ -1,16 +1,16 @@
 #include <cubos/core/settings.hpp>
 
-#include <cubos/engine/ecs/local_to_world.hpp>
-#include <cubos/engine/ecs/position.hpp>
-#include <cubos/engine/ecs/rotation.hpp>
-#include <cubos/engine/ecs/scale.hpp>
+#include <cubos/engine/transform/local_to_world.hpp>
 #include <cubos/engine/transform/plugin.hpp>
+#include <cubos/engine/transform/position.hpp>
+#include <cubos/engine/transform/rotation.hpp>
+#include <cubos/engine/transform/scale.hpp>
 
 using namespace cubos;
 
-static void applyTransform(core::ecs::Query<engine::ecs::LocalToWorld&, const engine::ecs::Position*,
-                                            const engine::ecs::Rotation*, const engine::ecs::Scale*>
-                               query)
+static void applyTransform(
+    core::ecs::Query<engine::LocalToWorld&, const engine::Position*, const engine::Rotation*, const engine::Scale*>
+        query)
 {
     for (auto [entity, localToWorld, position, rotation, scale] : query)
     {
@@ -32,10 +32,10 @@ static void applyTransform(core::ecs::Query<engine::ecs::LocalToWorld&, const en
 
 void cubos::engine::transformPlugin(Cubos& cubos)
 {
-    cubos.addComponent<ecs::Position>();
-    cubos.addComponent<ecs::Rotation>();
-    cubos.addComponent<ecs::Scale>();
-    cubos.addComponent<ecs::LocalToWorld>();
+    cubos.addComponent<Position>();
+    cubos.addComponent<Rotation>();
+    cubos.addComponent<Scale>();
+    cubos.addComponent<LocalToWorld>();
 
     cubos.system(applyTransform).tagged("cubos.transform.update");
 }
